fix: adding langchain to the Dockerfile
it's at the start as there is a clash in dependencies between langchain and the version of sqlalchemy Computer uses. 

The clash isn't a problem with the parts of langchain Computer uses, however updating langchain will revert the version of sqlalchemy. Hence doing langchain first so the poetry install updates the version.

Signed-off-by: Matt Thompson <thompsonson@gmail.com>